requirements.txt
altair
attrs
blinker
cachetools
certifi
charset-normalizer
click
colorama
contourpy
cycler
filelock
fonttools
fsspec
gitdb
GitPython
idna
Jinja2
jsonschema
jsonschema-specifications
kiwisolver
MarkupSafe
matplotlib
mpmath
narwhals
networkx
numpy
opencv-python
packaging
pandas
pillow
polars
protobuf
psutil
pyarrow
pydeck
pyparsing
python-dateutil
pytz
PyYAML
referencing
requests
rpds-py
scipy
setuptools
six
smmap
streamlit
sympy
tenacity
toml
torch
torchvision
tornado
typing_extensions
tzdata
ultralytics
ultralytics-thop
urllib3
watchdog
